pyproject.toml
[tool.poetry]
name = "dox-api"
version = "0.1.1"
description = ""
authors = ["Lalo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.237"
openai = "^0.27.8"
duckduckgo-search = "^3.8.3"
python-dotenv = "^1.0.0"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
gunicorn = "^21.2.0"
pymongo = {extras = ["srv"], version = "^4.4.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
